# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ollama Load Balancer API",
    "description": "REST API для интеллектуального балансировщика нагрузки Ollama с мониторингом ресурсов GPU/CPU/RAM/Disk и умным распределением запросов.\n\n## Особенности\n- Умная балансировка на основе доступных ресурсов\n- Model Affinity - направление запросов к серверам с загруженной моделью\n- Real-time мониторинг через WebSocket\n- Health Check - автоматическое обнаружение нерабочих бэкендов\n- Session Stickiness - сохранение сессии на одном сервере\n- Queue Manager - обработка перегрузок с очередью запросов\n\n## Аутентификация\nAPI использует токены для аутентификации. Токен передается в заголовке `X-API-Token`.\nMaster token (первый в списке) может создавать и отзывать другие токены.",
    "version": "1.0.0",
    "contact": {
      "name": "Ollama Load Balancer"
    },
    "license": {
      "name": "MIT License"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local development server"
    },
    {
      "url": "https://localhost:8443",
      "description": "Production server (HTTPS)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Custom server",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "Hostname or IP address"
        },
        "port": {
          "default": "8081",
          "description": "API port"
        }
      }
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/cluster": {
      "get": {
        "tags": ["Cluster"],
        "summary": "Get cluster status",
        "description": "Получение информации о состоянии кластера",
        "operationId": "getClusterStatus",
        "responses": {
          "200": {
            "description": "Cluster status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backends": {
      "get": {
        "tags": ["Backends"],
        "summary": "List all backends",
        "description": "Получение списка всех бэкендов",
        "operationId": "listBackends",
        "responses": {
          "200": {
            "description": "Backends list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "backends": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BackendInfo"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 2
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Backends"],
        "summary": "Add backend",
        "description": "Добавление нового бэкенда в кластер",
        "operationId": "addBackend",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBackendRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Backend added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "backend": {
                      "$ref": "#/components/schemas/BackendInfo"
                    },
                    "message": {
                      "type": "string",
                      "example": "Backend added successfully. Health check will run automatically."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Backend already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backends/{backend_id}": {
      "get": {
        "tags": ["Backends"],
        "summary": "Get backend info",
        "description": "Получение информации о конкретном бэкенде",
        "operationId": "getBackend",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "required": true,
            "description": "ID бэкенда",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backend info retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Backend not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Backends"],
        "summary": "Delete backend",
        "description": "Удаление бэкенда из кластера",
        "operationId": "deleteBackend",
        "parameters": [
          {
            "name": "backend_id",
            "in": "path",
            "required": true,
            "description": "ID бэкенда",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backend deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "id": {
                      "type": "string",
                      "example": "gpu-1"
                    },
                    "message": {
                      "type": "string",
                      "example": "Backend removed successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Backend not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List sessions",
        "description": "Получение списка активных сессий",
        "operationId": "listSessions",
        "responses": {
          "200": {
            "description": "Sessions list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionInfo"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Sessions"],
        "summary": "Clear all sessions",
        "description": "Очистка всех активных сессий",
        "operationId": "clearSessions",
        "responses": {
          "200": {
            "description": "All sessions cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "All sessions cleared"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models": {
      "get": {
        "tags": ["Models"],
        "summary": "List models",
        "description": "Получение списка запущенных моделей на всех бэкендах",
        "operationId": "listModels",
        "responses": {
          "200": {
            "description": "Models list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "models": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "example": {
                        "gpu-1": ["llama3.1:70b", "mistral:7b"],
                        "gpu-2": ["llama3.1:8b"]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/status": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get authentication status",
        "description": "Получение статуса системы аутентификации",
        "operationId": "getAuthStatus",
        "responses": {
          "200": {
            "description": "Authentication status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "API токен для аутентификации"
      }
    },
...
  }
}